Fix and rename 'onForBoarding' option
It was not negated before it was sent to the graphql query as 'omitNonBoarding', which means the opposite thing.

I also renamed it to 'includeNonBoarding' which makes more sense than 'onForBoarding' which sounds like a callback.
'onForBoarding' will still work, but a deprecation warning will be printed if the option is passed.