Fix requiring missing files in prod build
The types/ directory is outside the src/ dir, and its contents
will not be included in the transpiled bundle. This has been fine
so far, because it has only included types that are stripped away.

But with the addition of Scooters came some enums in this directory.
Enums are transpiled to JS objects that must be included in the code.
The SDK crashed because it could not find the necessary objects from
the types/ dir, because it was not included in the build.

I moved the types/scooters.ts file into src/scooters to fix this.
It's good to keep such related files together anyways, and no other
files were using these types. Only the most general and frequently
used types should be in the types/ folder.