Deprecate all OTP 1 / JP v2 related queries and types
The JourneyPlanner v2 API is deprecated and everybody should
move over to v3.

We will not migrate the existing SDK methods to the new
API version. We want to encourage users to create their own
queries, so that they will be able to leverage the full
flexibility of GraphQL.

There are powerful GraphQL libraries and frameworks like Apollo.
Combined with our generated types JourneyPlannerTypes or your own
generated types, you get both the full flexibility of GraphQL and
TypeScript coverage.